Proposed patches for sparsebitmap/dpt-dbms project.
The C++ patches mentioned in sparsebitmap/dpt-dbms issue 1 are included.

The content of 'sample projects'/'DPT with Python' is replaced.

Makefiles are provided to build the Python interface under Visual Studio,
MSys2, and mingw-w64 with Wine.

+++ source/dbapi/findspec_api.cpp
@@ -46,7 +46,7 @@ APIFindSpecification::APIFindSpecification(const FindOperator& o, const std::str
 {}
 
 APIFindSpecification::APIFindSpecification(APIBitMappedRecordSet& s)
-: target(new FindSpecification(FD_SET$, reinterpret_cast<unsigned _int32>(s.target)))
+: target(new FindSpecification(FD_SET$, reinterpret_cast<unsigned long _int32>(s.target)))
 {}
 
 APIFindSpecification::APIFindSpecification(const std::string& n, const FindOperator& o)
